Handle moving slideshow images better.
If an image has multiple files, liek p0, p1, p2, etc. then any json or ugoira files will have the same number as the last file. So like `12345_p4 - coolIMage.json`. But if I deleted p4, then it was leaving the json file behind. Now it checks to see if any files from teh image were moved already. If so, then move the extra files, otherwise delete it.

This can be improved I think by acutally scanning the destination folder for files `p0` to `p(n-1)`. That way it would work on leftover files as well and not just files moved at the same time as the parent.